Create function game() and fix bug in function playRound()
Function game() simulates 5 rounds of rock, paper, scissors. It asks the
player for their choice and then calls the playRound() function, passing
it the player's choice and the computer's randomized choice by calling
function computerPlay(). After every round, the result of the round is
displayed and the score of the winner is incremented. After five
rounds, the result of the game is displayed.

Bug in function playRound() caused any input that wasn't one of the three
valid inputs to be displayed as a tie. Now if the player provides an
invalid input, "Invalid Input!" will be displayed and no one scores for
that round will be incremented.